@@ -1,10 +1,10 @@
 name: mock_creator_generator
 description: Creates a mock for classes with a single call method making it easy for us to create those mocks automatically.
-version: 1.0.9
+version: 1.0.13
 homepage: https://github.com/atreeon/MockCreator
  
-environment:  
-  sdk: ">=2.10.0 <3.0.0"
+environment:
+  sdk: ">=2.12.0-0 <3.0.0"
   
 dependencies:  
   build: ^1.2.2
@@ -14,9 +14,9 @@ dependencies:
   quiver: ^2.0.1
   dartz: ^0.8.3
   basic_utils: ^2.4.8
-  analyzer: '>=0.36.0 <0.42.0'
+  analyzer: '>=0.41.0 <0.42.0'
   meta: ^1.1.8
-  typedef_for_fn_generator: ^1.0.1
+  typedef_for_fn_generator: ^1.0.2
 
 dev_dependencies:  
   build_runner: ^1.7.3
