@@ -2,10 +2,10 @@ import 'package:meta/meta.dart';
 import 'package:mock_creator_generator/src/helpers.dart';
 
 String createMockCreator({
-  @required String className,
-  @required String returnType,
-  @required List<NameType> paramsNormal,
-  @required List<NameType> paramsNamed,
+  required String className,
+  required String returnType,
+  required List<NameType> paramsNormal,
+  required List<NameType> paramsNamed,
 }) {
   var sb = StringBuffer();
 
